some styling merged in the mentor's side
