+++ client/src/App.css
@@ -3,7 +3,7 @@ body, html {
   width: 100%;
   height: 100%;
   background-color: rgb(233, 233, 247);
-  font-family: "Roboto";
+  font-family: 'Roboto', sans-serif;
 
 }
 
@@ -11,6 +11,7 @@ body, html {
 .App {
   width: 100%;
   height: 100%;
+  font-family: 'Roboto', sans-serif;
 }
