+++ client/src/Components/Mentors/MentorFeedback/GraduateElement.styles.js
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export const GraduateElementStyles = styled.div``;
