+++ client/src/index.css
@@ -8,6 +8,7 @@ body, html {
     width: 100%;
     height: 100%;
     overflow-x: hidden;
+    font-family: 'Roboto', sans-serif;
 }
 
 :root {
